+++ core/platform-services/gms/src/main/kotlin/org/michaelbel/movies/platform/gms/config/ConfigServiceImpl.kt
@@ -6,7 +6,7 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flowOf
 import org.michaelbel.movies.platform.main.config.ConfigService
 
-class ConfigServiceImpl @Inject constructor(
+internal class ConfigServiceImpl @Inject constructor(
     private val firebaseRemoteConfig: FirebaseRemoteConfig
 ): ConfigService {
