+++ core/platform-services/gms/src/main/kotlin/org/michaelbel/movies/platform/gms/googleapi/GoogleApiModule.kt
@@ -13,7 +13,5 @@ internal object GoogleApiModule {
 
     @Provides
     @Singleton
-    fun provideGoogleApiAvailability(): GoogleApiAvailability {
-        return GoogleApiAvailability.getInstance()
-    }
+    fun provideGoogleApiAvailability(): GoogleApiAvailability = GoogleApiAvailability.getInstance()
 }
