--- core/platform-services/gms/src/main/kotlin/org/michaelbel/movies/platform/gms/review/ReviewServiceModule.kt
@@ -1,17 +1,0 @@
-package org.michaelbel.movies.platform.gms.review
-
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-import org.michaelbel.movies.platform.main.review.ReviewService
-
-@Module
-@InstallIn(SingletonComponent::class)
-internal interface ReviewServiceModule {
-
-    @Binds
-    @Singleton
-    fun provideReviewService(service: ReviewServiceImpl): ReviewService
-}
